requirements.txt
autopep8==1.5.5
pdf2image==1.14.0
Pillow==8.1.0
pycodestyle==2.6.0
toml==0.10.2
